WIP: `read_sample` program
This is still a work in progress. Currently `read_sample` is able to read a generic CSV file
via the `read_csv` function. The `read_csv` function returns a dictionary, with all original
headers as keys and their corresponding values in lists. The `read_csv` function should be able
to cope with the request from Steven to "handle another data set with the same structure but
different scenario".