NEW: plot of reference data
Note that this plot is of the original data Steven gave me, which suffered from some GPS errors, e.g.,
if you plot the range data of the radar objects the data tends to be noisey at select times. I will upload
another plot soon with both the reference trajectory and closest matching radar object trajectory, as well
as velocity of each as a function of time.