TST: first unit tests
Adding the first round of unit tests. This first test makes sure that `proc_sample.read_csv` is returning np.ndarray for each data column in the CSV.

Note that tests had to be added in a bit of a hacky nature because my directory structure is not designed to be a standard Python package. Therefore relative imports aren't possible, and nosetests must be run from the top-level directory.